+++ CHANGELOG.md
@@ -15,3 +15,8 @@
 - Update skill calculations to reflect Rogue Trader with half of base stat for untrained
 - Update skills to only have untrained (half stats), trained (+0), expert (+10), and veteran (+20)
 - Fixed notes in items not updating and setup textEnrich for all note sections
+
+## 0.1.2
+
+- Update biography section to reflect rogue trader
+- Added Rank as a calculated field based on experience spent